Implemented Volley requests for manually input hosts
Added EditText widget and Ping button. Entering a
hostname into the EditText allows for the Ping
button to update a new TextView with the DeviceInfo
for a remote host. Volley makes the request and
NanoHTTPD replies. The LocationService returns the
callback with the expected DeviceInfo object rather
than a raw string (performs the JSON conversion
itself)