Send new peers all known peers, not just self
Sending only self to new peers leads to a not fully connected network
graph, due to NSD not being fully reliable or symmetric. Sending every
new peer a list of known peers (minus itself) allows for a fully
connected graph to be constructed, assuming that the newly joined peer
is discovered by any other peer through NSD and/or the newly joined peer
discovers any other peer through NSD.

Currently, the same POST request is sent multiple times - once for each
peer to notify the new peer about. This should probably be reworked to
send a single POST request with the payload as a JSON array, however,
the JSONArray org.json class and VolleyRequest for Arrays rather than
Objects are much less well defined and convenient to use. This should be
fixed as quickly as possible, once a better solution is found.

Asana ID #16131095080244 #16131095080234 #16131095080224 #16131095080238
ID #16302756717027 #16302756717031