moved BooleanExpressionTree classes to top level
for serialization/deserialization purposes, the name of a class is
included when it is serialized, if that class implements Node. this
allows us to create the correct concrete Node type on the other end. to
make this take up less space, and be easier to parse/read, the classes
are made top level. also, BooleanExpressionTree was becoming quite
cluttered anyway.